local-assistant/requirements.txt
altgraph==0.17.4
annotated-types==0.6.0
anyio==3.7.1
click==8.1.7
diskcache==5.6.3
exceptiongroup==1.2.0
fastapi==0.104.1
h11==0.14.0
idna==3.6
importlib-metadata==6.8.0
llama_cpp_python==0.2.20
macholib==1.16.3
numpy==1.26.2
packaging==23.2
pydantic==2.5.2
pydantic-settings==2.1.0
pydantic_core==2.14.5
pyinstaller==6.2.0
pyinstaller-hooks-contrib==2023.10
python-dotenv==1.0.0
sniffio==1.3.0
sse-starlette==1.8.2
starlette==0.27.0
starlette-context==0.3.6
typing_extensions==4.8.0
uvicorn==0.24.0.post1
zipp==3.17.0
